Moved literal
